finaly the transcribed text is inserted to the f editor
