+++ components/editor-components/index.js
@@ -1,0 +1,12 @@
+"use dom";
+
+export { default as PageLinkBlock } from "./PageLinkBlock";
+export { default as EditorToolbars } from "./EditorToolbars";
+export { default as KeyboardToolbarWrapper } from "./KeyboardToolbarWrapper";
+export { default as ConfirmDialog } from "./ConfirmDialog";
+export { default as BlockNoteEditor } from "./BlockNoteEditor";
+export {
+  insertTranscribedTextBlock,
+  insertPageLinkBlock,
+} from "./TranscriptionHandler";
+export { default as editorSchema } from "./editorSchema";
